+++ src/chat/chat.routes.js
@@ -2,6 +2,7 @@ import express from "express";
 import { createChatroom } from "./controller/chatroom.controller.js";
 import { showChatroom } from "./controller/chatroom.controller.js";
 import { deleteChatrooms } from "./controller/chatroom.controller.js";
+// import { sendMessages } from "./controller/chat.controller.js";
 
 const router = express.Router();
 
@@ -10,7 +11,10 @@ router.post("", createChatroom);
 
 // 채팅방 조회 API
 router.get("/:consumerId", showChatroom);
-export default router;
 
 // 채팅방 삭제 API
 router.delete("/delete", deleteChatrooms);
+
+// router.get("/:chatroomId/messages", showMessages);
+
+export default router;
